+++ bofire/benchmarks/api.py
@@ -21,13 +21,19 @@ from bofire.benchmarks.multi import (
 )
 from bofire.benchmarks.single import (
     Ackley,
+    Booth,
     Branin,
     Branin30,
+    CrossInTray,
+    Easom,
     Hartmann,
     Hartmann6plus,
     Himmelblau,
+    HolderTable,
     Multinormalpdfs,
     MultiTaskHimmelblau,
+    Rosenbrock,
+    SixHumpCamel,
 )
 
 
@@ -44,11 +50,17 @@ AnyMultiBenchmark = Union[
 ]
 AnySingleBenchmark = Union[
     Ackley,
+    Booth,
     Branin,
     Branin30,
+    CrossInTray,
+    Easom,
     Hartmann,
     Hartmann6plus,
     Himmelblau,
+    HolderTable,
     MultiTaskHimmelblau,
     Multinormalpdfs,
+    Rosenbrock,
+    SixHumpCamel,
 ]